+++ src/main/kotlin/nexters/admin/support/auth/JwtTokenProvider.kt
@@ -4,7 +4,7 @@ import io.jsonwebtoken.Claims
 import io.jsonwebtoken.JwtException
 import io.jsonwebtoken.Jwts
 import io.jsonwebtoken.security.Keys
-import nexters.admin.support.exception.UnauthenticatedException
+import nexters.admin.exception.UnauthenticatedException
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.stereotype.Component
 import java.nio.charset.StandardCharsets
