+++ src/test/kotlin/nexters/admin/domain/user/member/MemberTest.kt
@@ -3,7 +3,7 @@ package nexters.admin.domain.user.member
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import nexters.admin.createNewMember
-import nexters.admin.support.exception.UnauthenticatedException
+import nexters.admin.exception.UnauthenticatedException
 import nexters.admin.domain.user.Password
 import org.junit.jupiter.api.Test
