+++ src/test/kotlin/nexters/admin/support/auth/JwtTokenProviderTest.kt
@@ -2,7 +2,7 @@ package nexters.admin.support.auth
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
-import nexters.admin.support.exception.UnauthenticatedException
+import nexters.admin.exception.UnauthenticatedException
 import org.junit.jupiter.api.Test
 
 private const val VALID_SECRET_KEY = "testtesttesttesttesttesttesttest"
